requirements.txt
asgiref==3.8.1
asttokens==2.4.1
atomicwrites==1.4.1
attrs==22.1.0
beautifulsoup4==4.11.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==2.0.12
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.1
decorator==5.1.1
defusedxml==0.8.0rc2
Django==3.2.16
django-bootstrap5==22.2
django-cors-headers==4.5.0
django-extensions==3.2.3
django-filter==23.1
django-templated-mail==1.1.1
django_debug_toolbar==3.8.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
exceptiongroup==1.2.1
executing==2.0.1
Faker==12.0.1
flake8==5.0.4
idna==3.10
iniconfig==1.1.1
isort==5.13.2
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.4
MarkupSafe==3.0.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mixer==7.2.2
more-itertools==10.5.0
oauthlib==3.2.2
packaging==21.3
parso==0.8.4
pep8-naming==0.13.3
Pillow==9.3.0
pluggy==0.13.1
prompt_toolkit==3.0.47
pure-eval==0.2.2
py==1.11.0
pyaes==1.6.1
pycodestyle==2.9.1
pycparser==2.22
pydocstyle==6.3.0
pyflakes==2.5.0
pygame==2.5.2
Pygments==2.18.0
PyJWT==2.1.0
pyparsing==3.0.9
Pyrogram==2.0.106
PySocks==1.7.1
pyTelegramBotAPI==4.23.0
pytest==6.2.4
pytest-django==4.4.0
pytest-lazy-fixture==0.6.3
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python-dotenv==1.0.1
python3-openid==3.2.0
pytils==0.4.1
pytz==2022.6
requests==2.26.0
requests-oauthlib==2.0.0
six==1.16.0
snowballstemmer==2.2.0
social-auth-app-django==4.0.0
social-auth-core==4.5.4
soupsieve==2.5
sqlparse==0.4.3
stack-data==0.6.3
toml==0.10.2
tomli==2.0.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.20
wcwidth==0.2.13
webcolors==24.8.0
yapf==0.32.0
